test_automation/requirements_pytest.txt
# Core testing framework
pytest>=7.0.0
pytest-html>=3.1.0
pytest-xdist>=3.0.0
pytest-rerunfailures>=11.0.0
pytest-timeout>=2.1.0
pytest-cov>=4.0.0

# Appium and mobile testing
Appium-Python-Client>=3.0.0
selenium>=4.0.0

# Utilities
PyYAML>=6.0
requests>=2.28.0
Pillow>=9.0.0
allure-pytest>=2.12.0

# Development tools
black>=22.0.0
flake8>=5.0.0
pytest-mock>=3.8.0